+++ db/migrate/20260912100000_keep_the_column_order_of_an_answer.rb
@@ -1,0 +1,5 @@
+class KeepTheColumnOrderOfAnAnswer < ActiveRecord::Migration[8.1]
+  def up
+    change_column :omen_answers, :result, :json, default: [], null: false
+  end
+end